Workflows: Fixing Ubuntu Builds
Something has come up on the standard containers coming down with the
workflows and maturin is having a heck of a time working now.  Not
completely sure what's going on, but there's something definitely
broken.

Switching back to normal Ubuntu provided container image which puts us
in a slightly more expected position, and similar to Debian itself (and
eventually Fedora and EL).

Signed-off-by: John 'Warthog9' Hawley <jhawley@tenstorrent.com>